Correct compilation for 32 bit systems

+++ include/fast_ber/ber_types/Tag.hpp
@@ -3,7 +3,7 @@
 namespace fast_ber
 {
 
-using Tag = long;
+using Tag = long long;
 
 enum class UniversalTag : Tag
 {
